requirements.txt
annotated-types==0.5.0
numpy==1.25.2
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
typing_extensions==4.7.1
